Extract ManualInputManagerTestScene
