+++ osu.Framework.Tests/Visual/Input/TestSceneKeyBindingsGrid.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Testing;
+
+namespace osu.Framework.Tests.Visual.Input
+{
+    public class TestSceneKeyBindingsGrid : ManualInputManagerTestScene
+    {
+    }
+}
